# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smart Healthcare System API",
        "description": "API for Smart Healthcare System",
        "contact": {
            "email": "support@smart-healthcare.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/janji/ketersediaan-all": {
            "get": {
                "tags": [
                    "Appointment Management"
                ],
                "summary": "[PUBLIK] Cek semua ketersediaan dokter",
                "description": "Endpoint publik untuk melihat ketersediaan semua dokter untuk 7 hari ke depan tanpa parameter.",
                "operationId": "getAllKetersediaan",
                "responses": {
                    "200": {
                        "description": "Daftar ketersediaan semua dokter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id_dokter": {
                                                "type": "integer"
                                            },
                                            "nama_dokter": {
                                                "type": "string"
                                            },
                                            "spesialisasi": {
                                                "type": "string"
                                            },
                                            "biaya_konsultasi": {
                                                "type": "number"
                                            },
                                            "shift": {
                                                "type": "string"
                                            },
                                            "jadwal_ketersediaan": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "tanggal": {
                                                            "type": "string",
                                                            "format": "date"
                                                        },
                                                        "hari": {
                                                            "type": "string"
                                                        },
                                                        "jam_terisi": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string"
                                                            }
                                                        },
                                                        "shift": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
        "/janji/booking-cepat": {
            "post": {
                "tags": [
                    "Appointment Management"
                ],
                "summary": "[AMAN] Booking Janji Temu Cepat",
                "description": "Endpoint untuk membuat janji temu secara cepat dengan validasi slot tersedia",
                "operationId": "bookingCepat",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_dokter",
                                    "tanggal",
                                    "waktu_mulai"
                                ],
                                "properties": {
                                    "id_dokter": {
                                        "description": "ID Dokter",
                                        "type": "integer"
                                    },
                                    "tanggal": {
                                        "description": "Tanggal janji temu (YYYY-MM-DD)",
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "waktu_mulai": {
                                        "description": "Waktu mulai (HH:mm)",
                                        "type": "string"
                                    },
                                    "keluhan": {
                                        "description": "Keluhan pasien (opsional)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Janji temu berhasil dibooking",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Janji temu berhasil dibooking"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/JanjiTemu"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Slot sudah terisi atau error validasi"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/janji": {
            "get": {
                "tags": [
                    "Appointment Management"
                ],
                "summary": "[AMAN] Dapatkan semua janji temu",
                "description": "Endpoint untuk mendapatkan semua janji temu (Admin/Dokter bisa lihat semua, Pasien hanya lihat miliknya)",
                "operationId": "getAllJanjiTemu",
                "responses": {
                    "200": {
                        "description": "Daftar semua janji temu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/JanjiTemu"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/janji/search": {
            "get": {
                "tags": [
                    "Appointment Management"
                ],
                "summary": "[AMAN] Cari janji temu",
                "description": "Endpoint untuk mencari janji temu berdasarkan tanggal dan/atau nama dokter.",
                "operationId": "searchJanjiTemu",
                "parameters": [
                    {
                        "name": "tanggal",
                        "in": "query",
                        "description": "Filter berdasarkan tanggal (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-11-20"
                        }
                    },
                    {
                        "name": "nama_dokter",
                        "in": "query",
                        "description": "Filter berdasarkan nama dokter (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Raihan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar janji temu yang sesuai filter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SearchJanjiTemuResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Doctor": {
                "title": "Doctor (Model Lengkap)",
                "description": "Model Doctor lengkap dari database",
                "properties": {
                    "id_dokter": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "id_pengguna": {
                        "type": "integer"
                    },
                    "spesialisasi": {
                        "type": "string"
                    },
                    "no_lisensi": {
                        "type": "string"
                    },
                    "biaya_konsultasi": {
                        "type": "number",
                        "format": "float"
                    },
                    "shift": {
                        "type": "string",
                        "enum": [
                            "pagi",
                            "malam"
                        ]
                    },
                    "pengguna": {
                        "$ref": "#/components/schemas/Pengguna"
                    }
                },
                "type": "object"
            },
...
            "JanjiTemu": {
                "title": "Janji Temu (Respons)",
                "description": "Model Janji Temu (Respons)",
                "properties": {
                    "id_janji_temu": {
                        "type": "integer"
                    },
                    "id_pasien": {
                        "type": "integer"
                    },
                    "id_dokter": {
                        "type": "integer"
                    },
                    "tanggal_janji": {
                        "type": "string",
                        "format": "date"
                    },
                    "waktu_mulai": {
                        "type": "string",
                        "format": "time",
                        "example": "14:00"
                    },
                    "waktu_selesai": {
                        "type": "string",
                        "format": "time",
                        "example": "15:00"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "terjadwal",
                            "selesai",
                            "dibatalkan"
                        ]
                    },
                    "keluhan": {
                        "type": "string"
                    },
                    "pasien": {
                        "$ref": "#/components/schemas/Patient"
                    },
                    "dokter": {
                        "$ref": "#/components/schemas/Doctor"
                    }
                },
                "type": "object"
            },
...
            "SearchJanjiTemuResponse": {
                "title": "Search Janji Temu Response",
                "description": "Response untuk pencarian janji temu",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Pencarian janji temu berhasil"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/JanjiTemu"
                        }
                    }
                },
                "type": "object"
            },
            "Patient": {
                "title": "Patient",
                "description": "Model Profil Pasien",
                "properties": {
                    "id_pasien": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_pengguna": {
                        "type": "integer",
                        "example": 1
                    },
                    "tanggal_lahir": {
                        "type": "string",
                        "format": "date",
                        "example": "2003-12-01"
                    },
                    "golongan_darah": {
                        "type": "string",
                        "enum": [
                            "A",
                            "B",
                            "AB",
                            "O"
                        ]
                    },
                    "alamat": {
                        "type": "string",
                        "example": "jalan avengers"
                    },
                    "pengguna": {
                        "$ref": "#/components/schemas/Pengguna"
                    }
                },
                "type": "object"
            },
            "Pengguna": {
                "title": "Pengguna",
                "description": "Model Pengguna (data dasar)",
                "properties": {
                    "id_pengguna": {
                        "type": "integer",
                        "example": 1
                    },
                    "nama_lengkap": {
                        "type": "string",
                        "example": "Raihan Stark"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "raihanstark@gmail.com"
                    },
                    "no_telepon": {
                        "type": "string",
                        "example": "085156401611"
                    },
                    "role": {
                        "type": "string",
                        "example": "pasien"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}